+++ theories/bi/notation.v
@@ -93,30 +93,33 @@ Reserved Notation "P ={ E }=∗ Q"
   (at level 99, E at level 50, Q at level 200,
    format "'[' P  '/' ={ E }=∗  Q ']'").
 
-Reserved Notation "|={ E1 , E2 , E3 }▷=> Q"
+(** Step-taking fancy updates *)
+Reserved Notation "|={ E1 } [ E2 ]▷=> Q"
   (at level 99, E1, E2 at level 50, Q at level 200,
-   format "|={ E1 , E2 , E3 }▷=>  Q").
-Reserved Notation "P ={ E1 , E2 , E3 }▷=∗ Q"
+   format "|={ E1 } [ E2 ]▷=>  Q").
+Reserved Notation "P ={ E1 } [ E2 ]▷=∗ Q"
   (at level 99, E1, E2 at level 50, Q at level 200,
-   format "'[' P  '/' ={ E1 , E2 , E3 }▷=∗  Q ']'").
-Reserved Notation "|={ E1 , E2 }▷=> Q"
-  (at level 99, E1, E2 at level 50, Q at level 200,
-   format "|={ E1 , E2 }▷=>  Q").
-Reserved Notation "P ={ E1 , E2 }▷=∗ Q"
-  (at level 99, E1, E2 at level 50, Q at level 200,
-   format "'[' P  '/' ={ E1 , E2 }▷=∗  Q ']'").
+   format "'[' P  '/' ={ E1 } [ E2 ]▷=∗  Q ']'").
 Reserved Notation "|={ E }▷=> Q"
   (at level 99, E at level 50, Q at level 200,
    format "|={ E }▷=>  Q").
 Reserved Notation "P ={ E }▷=∗ Q"
   (at level 99, E at level 50, Q at level 200,
    format "'[' P  '/' ={ E }▷=∗  Q ']'").
-Reserved Notation "|={ E1 , E2 }▷=>^ n Q"
+
+(** Multi-step-taking fancy updates *)
+Reserved Notation "|={ E1 } [ E2 ]▷=>^ n Q"
   (at level 99, E1, E2 at level 50, n at level 9, Q at level 200,
-   format "|={ E1 , E2 }▷=>^ n  Q").
-Reserved Notation "P ={ E1 , E2 }▷=∗^ n Q"
+   format "|={ E1 } [ E2 ]▷=>^ n  Q").
+Reserved Notation "P ={ E1 } [ E2 ]▷=∗^ n Q"
   (at level 99, E1, E2 at level 50, n at level 9, Q at level 200,
-   format "P  ={ E1 , E2 }▷=∗^ n  Q").
+   format "P  ={ E1 } [ E2 ]▷=∗^ n  Q").
+Reserved Notation "|={ E }▷=>^ n Q"
+  (at level 99, E at level 50, n at level 9, Q at level 200,
+   format "|={ E }▷=>^ n  Q").
+Reserved Notation "P ={ E }▷=∗^ n Q"
+  (at level 99, E at level 50, n at level 9, Q at level 200,
+   format "P  ={ E }▷=∗^ n  Q").
 
 (** * Big Ops *)
 Reserved Notation "'[∗' 'list]' k ↦ x ∈ l , P"